Use CompletionCallbackExecutionSink on Android/iOS too
Follow the same approach as in https://github.com/dotnet/xharness/pull/906
Should fix https://github.com/dotnet/xharness/issues/924
